// lib/webportal/lambdas/export_guias_webportal/src/modules/documentos/dto/guide-filters.dto.d.ts
export declare class GuideFiltersDto {
    userId?: number;
    page?: number;
    limit?: number;
    sort?: string;
    order?: string;
    from?: Date;
    to?: Date;
    dateType?: string;
    guideNumber?: string;
    manifestNumber?: string;
    status?: string;
    locationType?: string;
    location?: string;
    participantType?: string;
    participant?: string;
    isSimplified?: boolean;
    marcas?: string;
    faltanteSobrante?: string;
    operationType?: string;
}